refactor: error handling utility functions refactor and relocation
